# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for my Laravel application.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Retrieve a list of all orders",
                "operationId": "6d89b6991269761984f5ea4ad68d91ad",
                "responses": {
                    "200": {
                        "description": "Order list retrieved successfully"
                    },
                    "404": {
                        "description": "Order list not found"
                    }
                }
            }
        },
        "/api/v1/order_register": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Store a new order",
                "operationId": "4f459bcf6e3e53f600b648bdbe380da9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shop_id",
                                    "manager_id",
                                    "order_id",
                                    "order_status",
                                    "order_date",
                                    "sender_id",
                                    "receiver_id",
                                    "order_weight",
                                    "sub_amount",
                                    "parcel_price",
                                    "code_price",
                                    "total_amount",
                                    "cash_status",
                                    "note",
                                    "order_type"
                                ],
                                "properties": {
                                    "shop_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "manager_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_status": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-07"
                                    },
                                    "sender_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "receiver_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_weight": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "sub_amount": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "parcel_price": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "code_price": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "total_amount": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "cash_status": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_type": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/order_update/{id}": {
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Update a order",
                "operationId": "57c3930ab693f10fc4789641a5b41817",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shop_id",
                                    "manager_id",
                                    "order_id",
                                    "order_status",
                                    "order_date",
                                    "sender_id",
                                    "receiver_id",
                                    "order_weight",
                                    "sub_amount",
                                    "parcel_price",
                                    "code_price",
                                    "total_amount",
                                    "cash_status",
                                    "note",
                                    "order_type"
                                ],
                                "properties": {
                                    "shop_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "manager_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_status": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-07"
                                    },
                                    "sender_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "receiver_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_weight": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "sub_amount": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "parcel_price": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "code_price": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "total_amount": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "cash_status": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "order_type": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/v1/order_destroy/{id}": {
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Delete a order",
                "operationId": "ac4377682782a517482941e6296c446f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Order to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/v1/order_show/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Retrieve a specific order",
                "operationId": "7dccc25a639f3cf1dad4e84e84990ea4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Order to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
    },
...
}